+++ R/normSpectra2D.R
@@ -6,6 +6,7 @@
 #' \code{\link{Spectra2D}} object.  There are currently two options:
 #' \itemize{
 #'   \item \code{"zero2one"} normalizes each 2D spectrum to a [0 \ldots{} 1] scale.
+#'   \item \code{"minusPlus"} normalizes each 2D spectrum to a [-1 \ldots{} 1] scale.
 #'   \item \code{"TotInt"} normalizes each 2D spectrum so that the total area is one.
 #' }
 #' 
@@ -21,7 +22,7 @@
 #' @keywords utilities
 #'
 #' @export
-#' @seealso \code{link{centscaleSpectra2D}} for another means of scaling.
+#' @seealso \code{\link{centscaleSpectra2D}} for another means of scaling.
 #'
 #' @examples
 #'
@@ -34,22 +35,40 @@ normSpectra2D <- function(spectra, method = "zero2one") {
 	
 	.chkArgs(mode = 21L)
 	chkSpectra(spectra)
-	ok <- c("zero2one", "TotInt")
+	ok <- c("zero2one", "TotInt", "minusPlus")
 	if (!method %in% ok) stop("Invalid method specified")
+	ns <- length(spectra$names)
 	
-# normalize each 2D spectrum to a [0...1] range:
+# normalize each 2D spectrum to a [0 ... 1] range:
 
 	if (method == "zero2one") {
-		for (i in 1:length(spectra$names)) {
+		rescale <- function (x, nx1, nx2, minx, maxx) {
+			nx = nx1 + (nx2 - nx1) * (x - minx)/(maxx - minx)
+            return(nx)
+        }
+		for (i in 1:ns) {
 			rMin <- min(spectra$data[[i]], na.rm = TRUE)
-			spectra$data[[i]] <- spectra$data[[i]] - rMin
 			rMax <- max(spectra$data[[i]], na.rm = TRUE)
-			spectra$data[[i]] <- spectra$data[[i]]/rMax
+			spectra$data[[i]] <- rescale(spectra$data[[i]], 0.0, 1.0, rMin, rMax)
+			}
+		}
+
+# normalize each 2D spectrum to a [-1 ... 1] range:
+	if (method == "minusPlus") {
+		rescale <- function (x, nx1, nx2, minx, maxx) {
+			nx = nx1 + (nx2 - nx1) * (x - minx)/(maxx - minx)
+            return(nx)
+        }
+
+		for (i in 1:ns) {
+			rMin <- min(spectra$data[[i]], na.rm = TRUE)
+			rMax <- max(spectra$data[[i]], na.rm = TRUE)
+			spectra$data[[i]] <- rescale(spectra$data[[i]], -1.0, 1.0, rMin, rMax)
 			}
 		}
 
 	if (method == "TotInt") {
-		for (i in 1:length(spectra$names)) {
+		for (i in 1:ns) {
 			spectra$data[[i]] <- spectra$data[[i]]/sum(spectra$data[[i]], na.rm = TRUE)
 			}
 		}